requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.8.0
click==8.1.3
Django==4.1.1
djangorestframework==3.14.0
drf-spectacular==0.24.2
inflection==0.5.1
iniconfig==1.1.1
jsonschema==4.16.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.3
pytest-django==4.5.2
pytz==2022.2.1
PyYAML==6.0
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
